Add script to extract posts from dumped joomla database
Include the results from the gallery section to use as a basis for the
new equivalent.

This looked like it would be surprisingly quick and simple after half an
hour of tinkering. However:
* fixing up URLs for links to the site archive was complicated. I'll
  need to know that stuff anyway in order to make the redirector behave,
  so it wasn't wasted time, but it was still a pain.
* Tweaking nested galleries took a while as they kept being ugly. I
  suspect there'll be transferable knowledge here for a native Jekyll
  gallery.

If it turns out later that there are other posts to duplicate, it should
be trivial, as all the posts make it through acceptably, not just the
gallery section.