Add srcset-generating plugin
This means we serve the smallest thumbnail that still looks good regardless of the user's display density.

So far, it only does the standalone srcset and can't generate srcset+sizes, so it's good when you know the px size of the <img> element, but not when you know the image will be the full width of the page but don't know the screen size.