Get rid of blog-like features
The RSS feed didn't work anyway - I think it needs a Jekyll plugin enabling, but we're not expecting to announce news, so there shouldn't be stuff for people to subscribe to.

The post list on the homepage also doesn't make sense unless we're running a blog or posting news. I just disabled it as otherwise it won't be discoverable if it ever becomes appropriate to turn it back on.